requirements.txt
# Core dependencies
pandas==2.0.3
pyarrow==14.0.2
openpyxl==3.1.3
cryptography==45.0.5

# Delta Lake connector
deltalake==0.15.0

# PySpark - ensure compatible with Delta Lake
pyspark==4.0.0

# Airflow MySQL provider (optional but useful)
apache-airflow-providers-mysql==5.5.0
